+++ config/default.conf
@@ -20,3 +20,19 @@ server	[
 	}
 
 ]
+server	[
+	listen	0.0.0.0:7000;
+	name	empty.com ;
+	
+	location	/	{
+		root	/home/samu/html/empty;
+		methods GET;
+		index	index.html favicon.ico;
+	}
+
+	location	/directory	{
+		methods GET;
+		index	index.html;
+		root	/home/samu/html/YoupiBanane;
+	}
+]
