+++ html/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Página de Ejemplo</title>
+    <title>Webserv test</title>
     <style>
         body {
             font-family: Arial, sans-serif;
@@ -38,17 +38,19 @@
 <body>
 
 <header>
-    <h1>Bienvenidos a Mi Página de Ejemplo</h1>
+    <h1>Bienvenidos a webserv</h1>
+	<h2>Por agserran-, bazuara y sacorder</h2>
 </header>
 
 <main>
-    <p>Este es un párrafo de ejemplo en mi página web. Aquí puedes poner cualquier tipo de contenido que desees compartir.</p>
-    <img src="https://via.placeholder.com/600x400" alt="Imagen de ejemplo">
-    <p>Visita <a href="https://www.example.com" target="_blank">este enlace</a> para más información.</p>
+	<p>Desarrollo:</p>
+    <img src="https://lh6.googleusercontent.com/proxy/CSZD9jzbl8NRFuceQV3VEjmZJmwqJ6kAMIQhcLsPy6vyYA-3LAqdnCK9YlrjswvIZEj-HYwS1uaZqvpJIA1-pzhh1aL4TwEFCH5g-xho2G4PIRN8ZTtS6liLHaXMPbMi6bENmg" alt="Equipo de desarrollo">
+    <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
+    <p>Visita <a href="/post.html">este enlace</a> para probar el cgi.</p>
 </main>
 
 <footer>
-    <p>&copy; 2024 Mi Página de Ejemplo. Todos los derechos reservados.</p>
+    <p>42 Madrid, 2024</p>
 </footer>
 
 </body>
